fix: `--outSAMunmapped` now correctly takes multiple args
